+++ lib/Extension/ReportExtension.php
@@ -65,6 +65,12 @@ class ReportExtension implements ExtensionInterface
     public const TAG_REPORT_RENDERER = 'report.renderer';
     public const PARAM_TEMPLATE_MAP = 'report.template_map';
     public const PARAM_TEMPLATE_PATHS = 'report.template_paths';
+
+    /**
+     * This configuration has been removed as it did not do anything. If this
+     * setting is given it will be ignored, and in 2.0 an exception will be
+     * raised used.
+     */
     public const PARAM_OUTPUT_DIR_HTML = 'report.html_output_dir';
 
     public function configure(OptionsResolver $resolver): void
@@ -76,7 +82,6 @@ class ReportExtension implements ExtensionInterface
             self::PARAM_TEMPLATE_PATHS => [
                 __DIR__ . '/../../templates'
             ],
-            self::PARAM_OUTPUT_DIR_HTML => '.phpbench/html',
             self::PARAM_TEMPLATE_MAP => [
                 'PhpBench\\Report\\Model' => 'html',
                 'PhpBench\\Expression\\Ast' => 'html/node'
@@ -85,7 +90,6 @@ class ReportExtension implements ExtensionInterface
 
         $resolver->setAllowedTypes(self::PARAM_TEMPLATE_MAP, ['array']);
         $resolver->setAllowedTypes(self::PARAM_TEMPLATE_PATHS, ['array']);
-        $resolver->setAllowedTypes(self::PARAM_OUTPUT_DIR_HTML, ['string']);
         $resolver->setAllowedTypes(self::PARAM_REPORTS, ['array']);
         $resolver->setAllowedTypes(self::PARAM_OUTPUTS, ['array']);
         $resolver->setAllowedTypes(self::PARAM_COMPONENTS, ['array']);
@@ -95,7 +99,6 @@ class ReportExtension implements ExtensionInterface
             self::PARAM_OUTPUTS => 'Report renderer configurations, see :doc:`report-renderers`',
             self::PARAM_TEMPLATE_MAP => 'Namespace prefix to template path map for object rendering',
             self::PARAM_TEMPLATE_PATHS => 'List of paths to load templates from',
-            self::PARAM_OUTPUT_DIR_HTML => 'Path in which to render HTML reports',
         ]);
     }