+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ------------------
 
 - Fix bug with misassigned labels in bar chart #931
+- Safely removed non-functioning config option `report.html_output_dir` #930
 
 1.1.2 (2021-09-25)
 ------------------