+++ CHANGELOG.md
@@ -15,6 +15,10 @@ Bug fixes:
 
 - Fix property (`.`) access precedence, it is now the same as `[` array
   access #928
+
+1.1.3 (2021-10-31)
+------------------
+
 - Fix bug with misassigned labels in bar chart #931
 - Safely removed non-functioning config option `report.html_output_dir` #930